migrate StashAction legacy to StashAction.test.js
